UDPClient properly reacts to buffer size = 0 and accepts callback = nullptr (such callback will be ignored and default_callback will be called, if it exists)

+++ common/src/udp_client.cpp
@@ -13,12 +13,29 @@
 common::UDPClient::UDPClient(uint16_t port, size_t input_buffer_size, std::unique_ptr<Callback> &&default_callback)
     : sendSocket(port), receiveSocket(port + 1), inputBuffer(input_buffer_size, 0), defaultCallback(std::move(default_callback)),
       receiverThread(&common::UDPClient::receiverThreadFunction, std::ref(*this))
-{}
+{
+    if(input_buffer_size == 0)
+    {
+        ExceptionInfo::warning("initialized UDPClient without buffer - you will never receive answer to what you send");
+        receiverThread.join();
+#ifndef NDEBUG
+        std::cerr << "receiverThread joined\n";
+#endif
+    }
+}
 
 common::UDPClient::UDPClient(common::UDPsocket &send_socket, common::UDPsocket &receive_socket, size_t input_buffer_size, std::unique_ptr<common::UDPClient::Callback> &&default_callback)
     : sendSocket(send_socket), receiveSocket(receive_socket), inputBuffer(input_buffer_size, 0), defaultCallback(std::move(default_callback)),
       receiverThread(&common::UDPClient::receiverThreadFunction, std::ref(*this))
 {
+    if(input_buffer_size == 0)
+    {
+        ExceptionInfo::warning("initialized UDPClient without buffer - you will never receive answer to what you send");
+        receiverThread.join();
+#ifndef NDEBUG
+        std::cerr << "receiverThread joined\n";
+#endif
+    }
     if(receiveSocket.getAddress().getPort() - sendSocket.getAddress().getPort() != 1)
         throw Exception("receiveSocket in UDPClient must be bound at port 1 greater than sendSocket");
 }
@@ -30,12 +47,15 @@ common::UDPClient::~UDPClient()
 #ifndef NDEBUG
     std::cerr << "UDPClient destructor... ";
 #endif
-    if(send(&magic, 1, receiveSocket.getAddress()) < 0)
-        ExceptionInfo::warning("CRITICAL ERROR - could not send kill message to receiver thread - the program might lock down");
-    receiverThread.join();
+    if(inputBuffer.size() != 0) // if it is equal 0, receiver thread does nothing and was already joined in constructor
+    {
+        if(send(&magic, 1, receiveSocket.getAddress()) < 0)
+            ExceptionInfo::warning("CRITICAL ERROR - could not send kill message to receiver thread - the program might lock down");
+        receiverThread.join();
 #ifndef NDEBUG
-    std::cerr << "receiverThread joined\n";
+        std::cerr << "receiverThread joined\n";
 #endif
+    }
 }
 
 int common::UDPClient::send(const char *data, size_t size, const Address &address) const
@@ -96,19 +116,10 @@ bool common::UDPClient::handleReceiveAndCheckIfEnd(const common::Address &addr, 
             callbackMap.erase(it);
         }
     }
-    if(found)
-    {
-#ifndef NDEBUG
-        if(callback == nullptr) // it can happen only due to programmer error, so in Release this check will be disabled
-        {
-            ExceptionInfo::warning("callback in callbackMap for address " + addr.toString() + " is nullptr");
-            return false;
-        }
-#endif
+    if(found && callback != nullptr)
         callback->callbackOnReceive(addr, message);
-    }
-    else if(message[0] == -1) // 'magic' message - end if it came from sendSocket of the same class instance
-        return addr.isLoopback(sendSocket.getAddress().getPort()); // the only place in which this function can return true
+    else if(message[0] == -1 && addr.isLoopback(sendSocket.getAddress().getPort()))
+        return true; // end=true if 'magic' message came from sendSocket of the same class instance
     else if(defaultCallback.get() != nullptr)
         defaultCallback->callbackOnReceive(addr, message);
     return false;
@@ -116,6 +127,8 @@ bool common::UDPClient::handleReceiveAndCheckIfEnd(const common::Address &addr, 
 
 void common::UDPClient::receiverThreadFunction()
 {
+    if(inputBuffer.size() == 0)
+        return;
     Address addr;
     bool receiverThreadShouldEnd = false;
     while(!receiverThreadShouldEnd)
